Друзья, мы официально выходим из режима пре-альфы и переходим к альфа-версии!
Основные изменения:
1.Создана первая версия интерфейса хранилища паролей, пока что почти не функционален
2.Добавлен новый класс Scroll. "Скроллинг" индекса длины [списка, словаря и т.д.] на единицу
2.Добавлена новая настройка для параметров проверки: теперь можно(как и раньше) шифровать звездочками свои пароли
3."Обычные символы" в настройках генератора переименованы в "Допустимые символы"

Технические изменения:
1.Обновлен виджет EraseWidget. Теперь его родитель - Frame, что позволило убрать у него несколько ненужных методов
2.Обновлены виджеты, имеющие аргумент master: теперь там стоит не "root", а "None" для гибкости
3.Все аргументы у виджетов, которые были ранее обязательны к указанию, теперь в этом не нуждаются(они заменены на "None")
4.Добавлен класс Stigma. В будущем будет служить для создания меток.

Исправления багов:
1.Исправлена инициализация
2.Оптимизированы некоторые строчки кода

+++ check_widg_commands.py
@@ -6,9 +6,11 @@ from generator_interface import *
 
 def check_password():
     password = CheckUI.check_password_entry.get()
-    if selection_letters(only_type_symbols(password, 'letters'), 'great_letters') == '' and allow_letter.get() and allow_gr_letter.get():
+    if selection_letters(only_type_symbols(password, 'letters'),
+                         'great_letters') == '' and allow_letter.get() and allow_gr_letter.get():
         result = 'Нет больших букв'
-    elif selection_letters(only_type_symbols(password, 'letters'), 'small_letters') == '' and allow_letter.get() and allow_sm_letter.get():
+    elif selection_letters(only_type_symbols(password, 'letters'),
+                           'small_letters') == '' and allow_letter.get() and allow_sm_letter.get():
         result = 'Нет маленьких букв'
     elif only_type_symbols(password, 'numbers') == '' and allow_number.get():
         result = 'В пароле нет цифр'
@@ -32,6 +34,7 @@ def delete_symbols():
         result = example_symbolbet.delete_symbols(list_of_symbols)
         messagebox.showinfo('Результат', result)
         automatically_update_list()
+
     if select_checksymbolbet_action.get() == 1:
         logic_delete_symbols(allowed_symbolbet)
     elif select_checksymbolbet_action.get() == 2:
@@ -44,11 +47,11 @@ def add_symbols():
         check_interface.add_symbols_entry.delete(0)
         result = example_symbolbet.add_symbols(list_of_symbols)
         messagebox.showinfo('Результат', result)
+
     if select_checksymbolbet_action.get() == 1:
         logic_add_symbols(allowed_symbolbet)
     elif select_checksymbolbet_action.get() == 2:
         logic_add_symbols(required_check_symbolbet)
-
 
 
 def stay_symbols():
@@ -57,6 +60,7 @@ def stay_symbols():
         check_interface.stay_symbols_entry.delete(0)
         result = example_symbolbet.stay_symbols(list_of_symbols)
         messagebox.showinfo("Результат", result)
+
     if select_checksymbolbet_action.get() == 1:
         logic_stay_symbols(allowed_symbolbet)
     elif select_checksymbolbet_action.get() == 2:
@@ -80,61 +84,64 @@ def get_range():
 
 
 def stay_letters():
-    if allow_letter.get():
+    if allow_letter.get() and check_interface.letter_checkbutton['state'] == NORMAL:
         allowed_symbolbet.delete_black_symbols(split(only_type_symbols(allowed_symbolbet.get_black(), 'letters')))
         check_interface.gr_letter_checkbutton['state'] = NORMAL
         check_interface.sm_letter_checkbutton['state'] = NORMAL
-    else:
+    elif not allow_letter.get() and check_interface.letter_checkbutton['state'] == NORMAL:
         allowed_symbolbet.add_black_symbols(split(only_type_symbols(allowed_symbolbet.get(), 'letters')))
         check_interface.gr_letter_checkbutton['state'] = DISABLED
         check_interface.sm_letter_checkbutton['state'] = DISABLED
 
 
 def stay_sm_letters():
-    if allow_sm_letter.get():
+    if allow_sm_letter.get() and check_interface.sm_letter_checkbutton['state'] == NORMAL:
         allowed_symbolbet.delete_black_symbols(
             split(selection_letters(only_type_symbols(allowed_symbolbet.get_black(), 'letters'), 'small_letters')))
         if allow_gr_letter.get():
             check_interface.letter_checkbutton['state'] = NORMAL
-    else:
+    elif not allow_sm_letter.get() and check_interface.sm_letter_checkbutton['state'] == NORMAL:
         allowed_symbolbet.add_black_symbols(
             split(selection_letters(only_type_symbols(allowed_symbolbet.get(), 'letters'), 'small_letters')))
         check_interface.letter_checkbutton['state'] = DISABLED
 
 
 def stay_gr_letters():
-    if allow_gr_letter.get():
+    if allow_gr_letter.get() and check_interface.gr_letter_checkbutton['state'] == NORMAL:
         allowed_symbolbet.delete_black_symbols(
             split(selection_letters(only_type_symbols(allowed_symbolbet.get_black(), 'letters'), 'great_letters')))
         if allow_sm_letter.get():
             check_interface.letter_checkbutton['state'] = NORMAL
-    else:
+    elif not allow_gr_letter.get() and check_interface.gr_letter_checkbutton['state'] == NORMAL:
         allowed_symbolbet.add_black_symbols(
             split(selection_letters(only_type_symbols(allowed_symbolbet.get(), 'letters'), 'great_letters')))
         check_interface.letter_checkbutton['state'] = DISABLED
 
 
 def stay_numbers():
-    if allow_number.get():
+    if allow_number.get() and check_interface.number_checkbutton['state'] == NORMAL:
         allowed_symbolbet.delete_black_symbols(split(only_type_symbols(allowed_symbolbet.get_black(), 'numbers')))
-    else:
+    elif not allow_number.get() and check_interface.number_checkbutton['state'] == NORMAL:
         allowed_symbolbet.add_black_symbols(split(only_type_symbols(allowed_symbolbet.get(), 'numbers')))
 
 
 def stay_other():
-    if allow_other.get():
+    if allow_other.get() and check_interface.other_checkbutton['state'] == NORMAL:
         allowed_symbolbet.delete_black_symbols(split(only_type_symbols(allowed_symbolbet.get_black(), 'other')))
-    else:
+    elif not allow_other.get() and check_interface.other_checkbutton['state'] == NORMAL:
         allowed_symbolbet.add_black_symbols(split(only_type_symbols(allowed_symbolbet.get(), 'other')))
+
 
 def automatically_update_list():
     if check_interface.get_list_entry.get() != '':
         get_list()
 
+
 def get_list():
     def logic_get_list(example_symbolbet):
         check_interface.get_list_entry.delete(0)
         check_interface.get_list_entry.insert(0, divide_word(example_symbolbet.get()))
+
     if select_checksymbolbet_action.get() == 1:
         logic_get_list(allowed_symbolbet)
     elif select_checksymbolbet_action.get() == 2:
@@ -145,6 +152,7 @@ def leave():
     check_interface.place_forget()
     mmenu_interface.place()
 
+
 def set_standart():
     allowed_symbolbet.delete_symbols(split(allowed_symbolbet.get()))
     allowed_symbolbet.add_symbols(split('abcdefghijklmnopqrstuvwxyzABCDEFGHIJKLMNOPQRSTUVWXYZ0123456789!@#$%&*'))
@@ -153,6 +161,14 @@ def set_standart():
     required_check_symbolbet.delete_symbols(split(required_check_symbolbet.get()))
     len_checkpas_range.set()
     automatically_update_list()
+    allow_number.set(1)
+    allow_letter.set(1)
+    allow_gr_letter.set(1)
+    allow_sm_letter.set(1)
+    allow_other.set(1)
+    check_interface.letter_checkbutton['state'] = NORMAL
+    check_interface.gr_letter_checkbutton['state'] = NORMAL
+    check_interface.sm_letter_checkbutton['state'] = NORMAL
 
 
 def set_generator_settings():
@@ -163,6 +179,22 @@ def set_generator_settings():
     required_check_symbolbet.add_symbols(split(required_symbolbet.get()))
     allowed_symbolbet.add_black_symbols(split(symbolbet.get_black()))
     len_checkpas_range.set(range_len_pas.get()[0], range_len_pas.get()[1])
+    allow_number.set(select_number.get())
+    allow_letter.set(select_letter.get())
+    allow_gr_letter.set(select_gr_letter.get())
+    allow_sm_letter.set(select_sm_letter.get())
+    allow_other.set(select_other.get())
+    stay_numbers()
+    stay_letters()
+    stay_gr_letters()
+    stay_sm_letters()
+    stay_other()
+
+def encode_passwords():
+    if select_encoding_passwords.get():
+        CheckUI.check_password_entry.show("*")
+    else:
+        CheckUI.check_password_entry.show()
 
 
 check_interface.check_password_button['command'] = check_password
@@ -180,3 +212,4 @@ check_interface.other_checkbutton['command'] = stay_other
 check_interface.set_default_button['command'] = set_standart
 check_interface.set_gen_settings_button['command'] = set_generator_settings
 check_interface.leave_button['command'] = leave
+check_interface.encode_passwords_button["command"] = encode_passwords
